Include PDFs in git repo
Since it was recommended that I make the PDFs available on github rather
than forcing readers to read raw LaTeX or compile it themselves.

This will be moot as soon as I switch to a new format for the
documentation that compiles to HTML that I can host on a github pages
site.